Refactor: Extract ToSql into its own module.
